Begin converting interface to pass integer types to back-end where appropriate.
git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@764 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1

+++ R/MxCharOrNumber.R
@@ -14,4 +14,4 @@
 #   limitations under the License.
 
 
-setClassUnion("MxCharOrNumber", c("character", "numeric"))
+setClassUnion("MxCharOrNumber", c("character", "integer"))
